Changes & Additions
Added Enemy Objects that will be added and animated inside of the empty "cave" spaces inside the tile map.

Also added hit detection between the enemies and tiles, and player and enemy.

Also made some changes to the way that the tiles, player and the enemy is translated and drawn on the canvas so that the hit detection works correctly.