new commit.
added a function that creates and updates a radius around the player based on the player's movement on the x & y axis. Tiles in the opposite direction of the player movement is dynamically removed while tiles in the direction that the player is moving towards are added.

Had some issues with hitting and destroying tiles that are sitting above the player on the y-axis so i updated the logic and made a boolean that checks the distance between the top of the player object, the bottom of the tile object and added a threshold that adds a maximum allowed distance for successfully hitting the target tile.